rename base to tokenize_base
